Consolidate string-truncation helpers into utils::strings
Three call sites in master maintained near-identical truncation
logic: Patroni REST error rendering (byte-bounded, no allocation),
admin/UI query previews (char-bounded, 120 chars, verbatim), and
per-eviction log lines (char-bounded, 80 chars, CR/LF→space,
ellipsis). After the eviction-logging consolidation in this PR the
char-based pair already lived together; merging the byte-based
helper into the same module finishes the job.

New module exposes truncate_bytes (zero-copy, char-boundary-safe),
truncate_chars (char-bounded String building block), preview_query
(120 chars verbatim), and truncate_query_for_log (80 chars,
CR/LF→space, ellipsis). The two query-render helpers share
truncate_chars internally. Sixteen unit tests cover the byte
boundary cases (patroni's previous truncate_str had none), char
truncation, multi-byte UTF-8, exact-limit boundary, and the
log-line normalisation rules.

Behaviour at call sites is unchanged. The deduplication is what
counts: the next author who needs to truncate text for a log,
preview, or wire body grep'ing the codebase will see one module
with documented limits instead of three inline copies drifting
apart.
