+++ src/server/mod.rs
@@ -15,9 +15,9 @@ mod server_backend;
 pub use parameters::ServerParameters;
 pub use prepared_statement_cache::{
     anon_len, anon_snapshot, gc_sweep_anon, gc_sweep_named, intern_query, named_len,
-    named_snapshot, now_monotonic_ms, preview_query, record_query_count, record_query_duration_us,
-    reset_interners_force, set_interner_worker_threads, truncate_query_for_log, AnonEntry,
-    CacheEntryKind, GcStats, NamedEntry, PreparedStatementCache,
+    named_snapshot, now_monotonic_ms, record_query_count, record_query_duration_us,
+    reset_interners_force, set_interner_worker_threads, AnonEntry, CacheEntryKind, GcStats,
+    NamedEntry, PreparedStatementCache,
 };
 
 #[cfg(test)]
