+++ src/utils/mod.rs
@@ -4,6 +4,7 @@ pub mod core_affinity;
 pub mod dashmap;
 pub mod debug_messages;
 pub mod rate_limit;
+pub mod strings;
 
 /// Format chrono::Duration in Go-style: `4m30s`, `1h30m`, `2d 4h30m`.
 pub fn format_duration(duration: &chrono::Duration) -> String {
